Add "Mine Enemies," discoverable AI colonies!
Um hello! I'm not a programmer or anything so i don't really understand a lot of whats going on with this file i'm editing here but i was hoping to request an addition to the mod/suggest an addition if possible! Its okay if not, I just thought it would be interesting. I feel like most people I know that enjoy this mod really love the building aspect of it, but once finished it feels like there isn't much left to do with the colony. The PVP system is a really cool idea and can be a lot of fun! however in my experience most gamers prefer to avoid the possibility of their hard-built colony getting destroyed, although they like doing the destroying themselves/combat experience... (funny how that works). This sparked an idea. I noticed that in the most recent version of minecolonies, mini abandoned colonies can be found like villages all over the world. I think it would be so fun and add a lot of adventure to the game if there was some way to create natrually spawned colonies in the world, and that through diplomacy the player might hve the option to ally themself with this colony or to perhaps gain enough favor to take over said colony as their own! Or the player might (with a gurgle of delight) conquer said colony subsequentially decide to add it to their kingdom or raze it to the ground. This would provide an endless drive to explore the world even in a single player setting/world and entertain beyond the foundation of a colony. However, if the player went to war with said colony but failed to conquer it quickly, then said colony might launch a raid on the player's own base with increased difficulty in comparrison to usual attacks from raiders. Another idea is that these auto-generated colonies i've coined as "mine enemies" (get it because they're my enemies) could grow proportionately to the size of the player's own existing colony. This might be a stretch, but I think it would make the game more lastingly fun. For example, if my town hall is level 1 still and I have a population of 10, the AI run colonies i encounter might also have a similar level and population, give or take a bit. If I have a maxed colony, and have conquered or allied myself with a few other colonies, expanding into a kingdom, then the colonies I make war with might bring allies of their own into the fray. Maybe this could cause waves of attacks during an enemy raid from a warring colony instead of one large wave, like in a pillager raid. Please consider adding mine enemies to your mod- that would really blow it out of the water for me and my friends! Thank you for all you do!